# docs/api/openapi.yaml
openapi: "3.0.3"
info:
  title: Open Jira API
  description: Open-source Jira clone REST API
  version: 1.0.0
servers:
  - url: /api/v1
    description: Base API path
paths:
...
  /custom-fields/{fieldID}/options:
    get:
      summary: List options for custom field
      operationId: listOptions
      tags: [CustomFields]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: fieldID
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: Option list }
    post:
      summary: Add option to custom field
      operationId: addOption
      tags: [CustomFields]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: fieldID
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [value]
              properties:
                value: { type: string }
      responses:
        '201': { description: Option created }
...
  /projects/{projectID}/timeline:
    get:
      summary: Get timeline (Gantt) data
      operationId: getTimeline
      tags: [Timeline]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: projectID
          in: path
          required: true
          schema: { type: string }
        - name: zoom
          in: query
          schema: { type: string, enum: [weeks, months, quarters] }
      responses:
        '200': { description: Timeline data }
  /projects/{projectID}/calendar:
    get:
      summary: Get calendar data
      operationId: getCalendar
      tags: [Calendar]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: projectID
          in: path
          required: true
          schema: { type: string }
        - name: year
          in: query
          required: true
          schema: { type: integer }
        - name: month
          in: query
          required: true
          schema: { type: integer }
      responses:
        '200': { description: Calendar data }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...